+++ lib/aws/cloud_hsm.ex
@@ -168,6 +168,10 @@ defmodule AWS.CloudHSM do
     request(client, "ModifyLunaClient", input, options)
   end
 
+  @spec request(map(), binary(), map(), list()) ::
+    {:ok, Poison.Parser.t | nil, Poison.Response.t} |
+    {:error, Poison.Parser.t} |
+    {:error, HTTPoison.Error.t}
   defp request(client, action, input, options) do
     client = %{client | service: "cloudhsm"}
     host = "cloudhsm.#{client.region}.#{client.endpoint}"
@@ -179,7 +183,7 @@ defmodule AWS.CloudHSM do
     headers = AWS.Request.sign_v4(client, "POST", url, headers, payload)
     case HTTPoison.post(url, payload, headers, options) do
       {:ok, response=%HTTPoison.Response{status_code: 200, body: ""}} ->
-        {:ok, response}
+        {:ok, nil, response}
       {:ok, response=%HTTPoison.Response{status_code: 200, body: body}} ->
         {:ok, Poison.Parser.parse!(body), response}
       {:ok, _response=%HTTPoison.Response{body: body}} ->