+++ lib/aws/code_commit.ex
@@ -133,6 +133,10 @@ defmodule AWS.CodeCommit do
     request(client, "UpdateRepositoryName", input, options)
   end
 
+  @spec request(map(), binary(), map(), list()) ::
+    {:ok, Poison.Parser.t | nil, Poison.Response.t} |
+    {:error, Poison.Parser.t} |
+    {:error, HTTPoison.Error.t}
   defp request(client, action, input, options) do
     client = %{client | service: "codecommit"}
     host = "codecommit.#{client.region}.#{client.endpoint}"
@@ -144,7 +148,7 @@ defmodule AWS.CodeCommit do
     headers = AWS.Request.sign_v4(client, "POST", url, headers, payload)
     case HTTPoison.post(url, payload, headers, options) do
       {:ok, response=%HTTPoison.Response{status_code: 200, body: ""}} ->
-        {:ok, response}
+        {:ok, nil, response}
       {:ok, response=%HTTPoison.Response{status_code: 200, body: body}} ->
         {:ok, Poison.Parser.parse!(body), response}
       {:ok, _response=%HTTPoison.Response{body: body}} ->