+++ lib/aws/dynamodb_streams.ex
@@ -103,6 +103,10 @@ defmodule AWS.DynamoDB.Streams do
     request(client, "ListStreams", input, options)
   end
 
+  @spec request(map(), binary(), map(), list()) ::
+    {:ok, Poison.Parser.t | nil, Poison.Response.t} |
+    {:error, Poison.Parser.t} |
+    {:error, HTTPoison.Error.t}
   defp request(client, action, input, options) do
     client = %{client | service: "streams.dynamodb"}
     host = "streams.dynamodb.#{client.region}.#{client.endpoint}"
@@ -114,7 +118,7 @@ defmodule AWS.DynamoDB.Streams do
     headers = AWS.Request.sign_v4(client, "POST", url, headers, payload)
     case HTTPoison.post(url, payload, headers, options) do
       {:ok, response=%HTTPoison.Response{status_code: 200, body: ""}} ->
-        {:ok, response}
+        {:ok, nil, response}
       {:ok, response=%HTTPoison.Response{status_code: 200, body: body}} ->
         {:ok, Poison.Parser.parse!(body), response}
       {:ok, _response=%HTTPoison.Response{body: body}} ->