+++ lib/aws/ecs.ex
@@ -32,7 +32,50 @@ defmodule AWS.ECS do
   Runs and maintains a desired number of tasks from a specified task
   definition. If the number of tasks running in a service drops below
   `desiredCount`, Amazon ECS spawns another instantiation of the task in the
-  specified cluster.
+  specified cluster. To update an existing service, see `UpdateService`.
+
+  You can optionally specify a deployment configuration for your service.
+  During a deployment (which is triggered by changing the task definition of
+  a service with an `UpdateService` operation), the service scheduler uses
+  the `minimumHealthyPercent` and `maximumPercent` parameters to determine
+  the deployment strategy.
+
+  If the `minimumHealthyPercent` is below 100%, the scheduler can ignore the
+  `desiredCount` temporarily during a deployment. For example, if your
+  service has a `desiredCount` of four tasks, a `minimumHealthyPercent` of
+  50% allows the scheduler to stop two existing tasks before starting two new
+  tasks. Tasks for services that *do not* use a load balancer are considered
+  healthy if they are in the `RUNNING` state; tasks for services that *do*
+  use a load balancer are considered healthy if they are in the `RUNNING`
+  state and the container instance it is hosted on is reported as healthy by
+  the load balancer. The default value for `minimumHealthyPercent` is 50% in
+  the console and 100% for the AWS CLI, the AWS SDKs, and the APIs.
+
+  The `maximumPercent` parameter represents an upper limit on the number of
+  running tasks during a deployment, which enables you to define the
+  deployment batch size. For example, if your service has a `desiredCount` of
+  four tasks, a `maximumPercent` value of 200% starts four new tasks before
+  stopping the four older tasks (provided that the cluster resources required
+  to do this are available). The default value for `maximumPercent` is 200%.
+
+  When the service scheduler launches new tasks, it attempts to balance them
+  across the Availability Zones in your cluster with the following logic:
+
+  <ul> <li> Determine which of the container instances in your cluster can
+  support your service's task definition (for example, they have the required
+  CPU, memory, ports, and container instance attributes).
+
+  </li> <li> Sort the valid container instances by the fewest number of
+  running tasks for this service in the same Availability Zone as the
+  instance. For example, if zone A has one running service task and zones B
+  and C each have zero, valid container instances in either zone B or C are
+  considered optimal for placement.
+
+  </li> <li> Place the new service task on a valid container instance in an
+  optimal Availability Zone (based on the previous steps), favoring container
+  instances with the fewest number of running tasks for this service.
+
+  </li> </ul>
   """
   def create_service(client, input, options \\ []) do
     request(client, "CreateService", input, options)
@@ -49,7 +92,23 @@ defmodule AWS.ECS do
   end
 
   @doc """
-  Deletes a specified service within a cluster.
+  Deletes a specified service within a cluster. You can delete a service if
+  you have no running tasks in it and the desired task count is zero. If the
+  service is actively maintaining tasks, you cannot delete it, and you must
+  update the service to a desired task count of zero. For more information,
+  see `UpdateService`.
+
+  <note> When you delete a service, if there are still running tasks that
+  require cleanup, the service status moves from `ACTIVE` to `DRAINING`, and
+  the service is no longer visible in the console or in `ListServices` API
+  operations. After the tasks have stopped, then the service status moves
+  from `DRAINING` to `INACTIVE`. Services in the `DRAINING` or `INACTIVE`
+  status can still be viewed with `DescribeServices` API operations; however,
+  in the future, `INACTIVE` services may be cleaned up and purged from Amazon
+  ECS record keeping, and `DescribeServices` API operations on those services
+  will return a `ServiceNotFoundException` error.
+
+  </note>
   """
   def delete_service(client, input, options \\ []) do
     request(client, "DeleteService", input, options)
@@ -303,34 +362,73 @@ defmodule AWS.ECS do
   end
 
   @doc """
-  Modify the desired count or task definition used in a service.
+  Modifies the desired count, deployment configuration, or task definition
+  used in a service.
 
   You can add to or subtract from the number of instantiations of a task
   definition in a service by specifying the cluster that the service is
   running in and a new `desiredCount` parameter.
 
   You can use `UpdateService` to modify your task definition and deploy a new
-  version of your service, one task at a time. If you modify the task
-  definition with `UpdateService`, Amazon ECS spawns a task with the new
-  version of the task definition and then stops an old task after the new
-  version is running. Because `UpdateService` starts a new version of the
-  task before stopping an old version, your cluster must have capacity to
-  support one more instantiation of the task when `UpdateService` is run. If
-  your cluster cannot support another instantiation of the task used in your
-  service, you can reduce the desired count of your service by one before
-  modifying the task definition.
+  version of your service.
 
-  When `UpdateService` replaces a task during an update, the equivalent of
+  You can also update the deployment configuration of a service. When a
+  deployment is triggered by updating the task definition of a service, the
+  service scheduler uses the deployment configuration parameters,
+  `minimumHealthyPercent` and `maximumPercent`, to determine the deployment
+  strategy.
+
+  If the `minimumHealthyPercent` is below 100%, the scheduler can ignore the
+  `desiredCount` temporarily during a deployment. For example, if your
+  service has a `desiredCount` of four tasks, a `minimumHealthyPercent` of
+  50% allows the scheduler to stop two existing tasks before starting two new
+  tasks. Tasks for services that *do not* use a load balancer are considered
+  healthy if they are in the `RUNNING` state; tasks for services that *do*
+  use a load balancer are considered healthy if they are in the `RUNNING`
+  state and the container instance it is hosted on is reported as healthy by
+  the load balancer.
+
+  The `maximumPercent` parameter represents an upper limit on the number of
+  running tasks during a deployment, which enables you to define the
+  deployment batch size. For example, if your service has a `desiredCount` of
+  four tasks, a `maximumPercent` value of 200% starts four new tasks before
+  stopping the four older tasks (provided that the cluster resources required
+  to do this are available).
+
+  When `UpdateService` stops a task during a deployment, the equivalent of
   `docker stop` is issued to the containers running in the task. This results
   in a `SIGTERM` and a 30-second timeout, after which `SIGKILL` is sent and
   the containers are forcibly stopped. If the container handles the `SIGTERM`
   gracefully and exits within 30 seconds from receiving it, no `SIGKILL` is
   sent.
+
+  When the service scheduler launches new tasks, it attempts to balance them
+  across the Availability Zones in your cluster with the following logic:
+
+  <ul> <li> Determine which of the container instances in your cluster can
+  support your service's task definition (for example, they have the required
+  CPU, memory, ports, and container instance attributes).
+
+  </li> <li> Sort the valid container instances by the fewest number of
+  running tasks for this service in the same Availability Zone as the
+  instance. For example, if zone A has one running service task and zones B
+  and C each have zero, valid container instances in either zone B or C are
+  considered optimal for placement.
+
+  </li> <li> Place the new service task on a valid container instance in an
+  optimal Availability Zone (based on the previous steps), favoring container
+  instances with the fewest number of running tasks for this service.
+
+  </li> </ul>
   """
   def update_service(client, input, options \\ []) do
     request(client, "UpdateService", input, options)
   end
 
+  @spec request(map(), binary(), map(), list()) ::
+    {:ok, Poison.Parser.t | nil, Poison.Response.t} |
+    {:error, Poison.Parser.t} |
+    {:error, HTTPoison.Error.t}
   defp request(client, action, input, options) do
     client = %{client | service: "ecs"}
     host = "ecs.#{client.region}.#{client.endpoint}"
@@ -342,7 +440,7 @@ defmodule AWS.ECS do
     headers = AWS.Request.sign_v4(client, "POST", url, headers, payload)
     case HTTPoison.post(url, payload, headers, options) do
       {:ok, response=%HTTPoison.Response{status_code: 200, body: ""}} ->
-        {:ok, response}
+        {:ok, nil, response}
       {:ok, response=%HTTPoison.Response{status_code: 200, body: body}} ->
         {:ok, Poison.Parser.parse!(body), response}
       {:ok, _response=%HTTPoison.Response{body: body}} ->